Use minimal hosting
Switch to minimal hosting for the API.
